local-registry: configure image signature and encryption too
