+++ tasks/eval/__init__.py
@@ -1,6 +1,7 @@
 from invoke import Collection
 
 from . import image_pull
+from . import images
 from . import mem_size
 from . import startup
 from . import vm_detail
@@ -9,6 +10,7 @@ from . import xput_detail
 
 ns = Collection(
     image_pull,
+    images,
     mem_size,
     startup,
     vm_detail,
